Move modules depending on frontier types into frontier hardfork
